fix: 데이터 컴포넌트에서 페이지로 옮김
