+++ components/agenda/teamDetail/TeamInfo.tsx
@@ -1,46 +1,17 @@
-import { useState } from 'react';
 import { TeamDetailProps } from 'types/aganda/TeamDetailTypes';
-import { TeamStatus, Coalition, AgendaLocation } from 'constants/agenda/agenda';
+import { Coalition } from 'constants/agenda/agenda';
 import colorstyles from 'styles/agenda/coalition.module.scss';
 import styles from 'styles/agenda/TeamDetail/TeamInfo.module.scss';
 import { ShareBtn } from '../button/Buttons';
 
 //인원 받아올려면 Team Max 인원 받아와야 함
-const TeamInfo = () => {
-  const [teamDetail, setTeamDetail] = useState<TeamDetailProps>({
-    teamName: '팀이름',
-    teamLeaderIntraId: 'leader',
-    teamStatus: TeamStatus.OPEN, //e
-    teamLocation: AgendaLocation.MIX, //e
-    teamContent: '우리팀이세계최강팀',
-    teamMates: [
-      {
-        intraId: 'leader',
-        coalition: Coalition.GUN, //e
-      },
-      {
-        intraId: 'member1',
-        coalition: Coalition.SPRING,
-      },
-      {
-        intraId: 'member1',
-        coalition: Coalition.SPRING,
-      },
-      {
-        intraId: 'member1',
-        coalition: Coalition.SPRING,
-      },
-      {
-        intraId: 'member2',
-        coalition: Coalition.SUMMER,
-      },
-    ],
-  });
-
-  if (teamDetail.teamStatus === TeamStatus.CANCEL) {
-    setTeamDetail({ ...teamDetail, teamStatus: TeamStatus.CONFIRM });
-  }
-
+const TeamInfo = ({
+  teamDetail,
+  shareTeamInfo,
+}: {
+  teamDetail: TeamDetailProps;
+  shareTeamInfo: () => void;
+}) => {
   const colorMapping: { [key: string]: string } = {
     [Coalition.GUN]: colorstyles.bg_gun,
     [Coalition.GON]: colorstyles.bg_gon,
@@ -52,11 +23,6 @@ const TeamInfo = () => {
     [Coalition.WINTER]: colorstyles.bg_winter,
     [Coalition.OTHER]: colorstyles.bg_default,
   };
-
-  const shareTeamInfo = () => {
-    alert('공유하기');
-  };
-  console.log(colorMapping[teamDetail.teamMates[0].coalition]);
 
   return (
     <div className={styles.container}>
